requirements.txt
streamlit
pandas
Pillow
streamlit-option-menu
plotly
streamlit-lottie